services/visualization/FinalVisualization/requirements.txt
numpy
pandas
requests
flask
flask_executor
pyecharts
cassandra-driver
flask_apscheduler